Rework custom event naming to match Fabric API
