+++ src/main/java/grojdg/harderstats/mixin/ServerPlayerEntityMixin.java
@@ -1,6 +1,6 @@
 package grojdg.harderstats.mixin;
 
-import grojdg.harderstats.callbacks.*;
+import grojdg.harderstats.events.*;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.server.network.ServerPlayerEntity;
