+++ src/main/kotlin/grojdg/harderstats/HarderStats.kt
@@ -1,8 +1,11 @@
 package grojdg.harderstats
 
 import grojdg.harderstats.accessors.TickAccess
-import grojdg.harderstats.callbacks.*
-import grojdg.harderstats.events.*
+import grojdg.harderstats.events.LivingEntityEvents
+import grojdg.harderstats.events.PlayerManagerEvents
+import grojdg.harderstats.events.ServerPlayerEntityEvents
+import grojdg.harderstats.events.WaterEvents
+import grojdg.harderstats.listeners.*
 import net.fabricmc.api.ModInitializer
 import net.fabricmc.fabric.api.entity.event.v1.ServerEntityCombatEvents
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents
@@ -17,14 +20,14 @@ object HarderStats : ModInitializer {
 	override fun onInitialize() {
 		LOGGER.info("Harderstats initializing");
 
-		WaterCallback.event.register(WaterEvents::onUpdateSubmergedInWater)
-		AfterDamage.event.register(ServerPlayerEntityEvents::onDamage)
-		AfterDeath.event.register(ServerPlayerEntityEvents::onDeath)	// TODO swap with fabric api official AfterDeath event
-		AfterExperienceGain.event.register(ServerPlayerEntityEvents::onExperienceGain)
-		OnMoveToWorld.event.register(ServerPlayerEntityEvents::onMoveToWorld)
-		ServerEntityCombatEvents.AFTER_KILLED_OTHER_ENTITY.register(EntityEvents::onKilledEntity)
-		OnFoodEaten.event.register(FoodEvents::onEatFood)
-		AfterPlayerConnect.event.register(PlayerManagerEvents::afterPlayerConnect)
+		WaterEvents.AFTER_WATER_SUBMERSION_STATE_UPDATE.register(WaterListeners::onUpdateSubmergedInWater)
+		ServerPlayerEntityEvents.AFTER_DAMAGE.register(ServerPlayerEntityListeners::onDamage)
+		ServerPlayerEntityEvents.AFTER_DEATH.register(ServerPlayerEntityListeners::onDeath)	// TODO swap with fabric api official AfterDeath event
+		ServerPlayerEntityEvents.AFTER_EXPERIENCE_GAIN.register(ServerPlayerEntityListeners::onExperienceGain)
+		ServerPlayerEntityEvents.ON_MOVE_TO_WORLD.register(ServerPlayerEntityListeners::onMoveToWorld)
+		ServerEntityCombatEvents.AFTER_KILLED_OTHER_ENTITY.register(CombatListeners::onKilledEntity)
+		LivingEntityEvents.ON_FOOD_EATEN.register(FoodListeners::onEatFood)
+		PlayerManagerEvents.AFTER_PLAYER_CONNECT.register(PlayerManagerListeners::afterPlayerConnect)
 
 		// set our timer
 		ServerLifecycleEvents.SERVER_STARTED.register(ServerStarted { server: MinecraftServer -> (server as TickAccess).harderstats_setTimer(InfoReceptionService.dispatchTime) })
