+++ src/main/kotlin/grojdg/harderstats/events/ServerPlayerEntityEvents.kt
@@ -1,48 +1,87 @@
 package grojdg.harderstats.events
 
-import grojdg.harderstats.HarderStats
-import grojdg.harderstats.InfoReceptionService
+import net.fabricmc.fabric.api.event.Event
+import net.fabricmc.fabric.api.event.EventFactory
 import net.minecraft.entity.damage.DamageSource
 import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.server.world.ServerWorld
-import net.minecraft.world.World
 
 object ServerPlayerEntityEvents {
-    fun onDamage(source: DamageSource, amount: Float, player: ServerPlayerEntity) {
-        if (source.name != "hih" && source.type.msgId() != "outOfWorld") {
-            InfoReceptionService.updateDamageTaken(player.gameProfile.id, amount.toInt())
+    fun interface AfterDamage {
+        fun interact(source: DamageSource, amount: Float, player: ServerPlayerEntity)
+    }
+
+    val AFTER_DAMAGE: Event<AfterDamage> = EventFactory.createArrayBacked(AfterDamage::class.java
+    ) { listeners ->
+        AfterDamage { source, amount, player ->
+            for (listener in listeners) {
+                listener.interact(source, amount, player)
+            }
         }
     }
 
-    fun onDeath(source: DamageSource, player: ServerPlayerEntity) {
-        // this is the type we kill with manually in HeartOfHearts
-        if (source.type.msgId() != "outOfWorld") {
-            val entityName = if (source.source == null) "null" else source.source!!.name.string
-            InfoReceptionService.dispatchWorldDeath(
-                    player.gameProfile.id,
-                    entityName,
-                    source.type.msgId()
-            )
+    fun interface AfterDeath {
+        fun interact(source: DamageSource, player: ServerPlayerEntity)
+    }
+
+    val AFTER_DEATH: Event<AfterDeath> = EventFactory.createArrayBacked(AfterDeath::class.java
+    ) { listeners ->
+        AfterDeath { source, player ->
+            for (listener in listeners) {
+                listener.interact(source, player)
+            }
         }
     }
 
-    fun onExperienceGain(experience: Int, player: ServerPlayerEntity) {
-        InfoReceptionService.updateExperienceGained(player.uuid, experience)
+    fun interface AfterExperienceGain {
+        fun interact(experience: Int, player: ServerPlayerEntity)
     }
 
-    // we need to reset timers on disconnect
-    fun onPlayerDisconnect(player: ServerPlayerEntity) {
-        InfoReceptionService.setIsInWater(player.uuid, false)
-        InfoReceptionService.setIsInNether(player.uuid, false)
-    }
-
-    fun onMoveToWorld(destination: ServerWorld, player: ServerPlayerEntity) {
-        HarderStats.LOGGER.info("Player is being moved into a world")
-        if (destination.registryKey === World.NETHER) {
-            InfoReceptionService.setIsInNether(player.uuid, true)
+    val AFTER_EXPERIENCE_GAIN: Event<AfterExperienceGain> = EventFactory.createArrayBacked(AfterExperienceGain::class.java
+    ) { listeners ->
+        AfterExperienceGain { experience, player ->
+            for (listener in listeners) {
+                listener.interact(experience, player)
+            }
         }
-        if (destination.registryKey === World.OVERWORLD) {
-            InfoReceptionService.setIsInNether(player.uuid, false)
+    }
+
+    fun interface AfterPlayerDisconnect {
+        fun interact(player: ServerPlayerEntity?)
+    }
+
+    val AFTER_PLAYER_DISCONNECT: Event<AfterPlayerDisconnect> = EventFactory.createArrayBacked(AfterPlayerDisconnect::class.java
+    ) { listeners ->
+        AfterPlayerDisconnect { player ->
+            for (listener in listeners) {
+                listener.interact(player)
+            }
+        }
+    }
+
+    fun interface OnMoveToWorld {
+        fun interact(destination: ServerWorld, player: ServerPlayerEntity)
+    }
+
+    val ON_MOVE_TO_WORLD: Event<OnMoveToWorld> = EventFactory.createArrayBacked(OnMoveToWorld::class.java
+    ) { listeners ->
+        OnMoveToWorld { world, player ->
+            for (listener in listeners) {
+                listener.interact(world, player)
+            }
+        }
+    }
+
+    fun interface OnConstructPlayer {
+        fun interact(world: ServerWorld, player: ServerPlayerEntity)
+    }
+
+    val ON_CONSTRUCT_PLAYER: Event<OnConstructPlayer> = EventFactory.createArrayBacked(OnConstructPlayer::class.java
+    ) { listeners ->
+        OnConstructPlayer { world, player ->
+            for (listener in listeners) {
+                listener.interact(world, player)
+            }
         }
     }
 }
