+++ src/main/kotlin/grojdg/harderstats/events/WaterEvents.kt
@@ -1,10 +1,21 @@
 package grojdg.harderstats.events
 
-import grojdg.harderstats.InfoReceptionService
+import net.fabricmc.fabric.api.event.Event
+import net.fabricmc.fabric.api.event.EventFactory
 import net.minecraft.entity.player.PlayerEntity
 
 object WaterEvents {
-    fun onUpdateSubmergedInWater(player: PlayerEntity) {
-        InfoReceptionService.setIsInWater(player.uuid, player.isWet)
+    fun interface AfterWaterSubmersionStateUpdate {
+        fun interact(player: PlayerEntity)
     }
+
+    val AFTER_WATER_SUBMERSION_STATE_UPDATE: Event<AfterWaterSubmersionStateUpdate> = EventFactory.createArrayBacked(AfterWaterSubmersionStateUpdate::class.java
+    ) { listeners ->
+        AfterWaterSubmersionStateUpdate { player: PlayerEntity ->
+            for (listener in listeners) {
+                listener.interact(player)
+            }
+        }
+    }
+
 }
