+++ src/main/kotlin/grojdg/harderstats/listeners/CombatListeners.kt
@@ -1,4 +1,4 @@
-package grojdg.harderstats.events
+package grojdg.harderstats.listeners
 
 import grojdg.harderstats.InfoReceptionService
 import net.minecraft.entity.Entity
@@ -6,7 +6,7 @@ import net.minecraft.entity.LivingEntity
 import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.server.world.ServerWorld
 
-object EntityEvents {
+object CombatListeners {
     fun onKilledEntity(world: ServerWorld?, entity: Entity, killedEntity: LivingEntity) {
         if (entity.isPlayer && !killedEntity.isPlayer) {
             InfoReceptionService.updateMobsKilled((entity as ServerPlayerEntity).gameProfile.id, 1)
